[JSC] Extract MicrotaskQueue https://bugs.webkit.org/show_bug.cgi?id=289074 rdar://146088353
Reviewed by Yijia Huang.

This patch extracts JSC's MicrotaskQueue to prepare to extend it to
fully cover WebCore use case. Right now it is only used for
JavaScriptCore framework etc.

1. This patch extends MicrotaskQueue to accept multiple instances. We
   have a linked list in VM and managing all the instances of
   MicrotaskQueue associated with this VM.
2. Move MicrotaskQueue to its own file.

* Source/JavaScriptCore/CMakeLists.txt:
* Source/JavaScriptCore/JavaScriptCore.xcodeproj/project.pbxproj:
* Source/JavaScriptCore/Sources.txt:
* Source/JavaScriptCore/runtime/MicrotaskQueue.cpp: Added.
(JSC::QueuedTask::run):
(JSC::MicrotaskQueue::create):
(JSC::MicrotaskQueue::visitAggregateImpl):
* Source/JavaScriptCore/runtime/MicrotaskQueue.h: Added.
(JSC::QueuedTask::QueuedTask):
(JSC::QueuedTask::identifier const):
* Source/JavaScriptCore/runtime/VM.cpp:
(JSC::VM::VM):
(JSC::VM::queueMicrotask):
(JSC::VM::drainMicrotasks):
(JSC::VM::beginMarking):
(JSC::VM::visitAggregateImpl):
(JSC::QueuedTask::run): Deleted.
(JSC::MicrotaskQueue::visitAggregateImpl): Deleted.
* Source/JavaScriptCore/runtime/VM.h:
(JSC::QueuedTask::QueuedTask): Deleted.
(JSC::QueuedTask::identifier const): Deleted.
(JSC::MicrotaskQueue::dequeue): Deleted.
(JSC::MicrotaskQueue::enqueue): Deleted.
(JSC::MicrotaskQueue::isEmpty const): Deleted.
(JSC::MicrotaskQueue::size const): Deleted.
(JSC::MicrotaskQueue::clear): Deleted.
(JSC::MicrotaskQueue::beginMarking): Deleted.
* Tools/TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/291566@main
